+++ src/api/csw/getRecordById.js
@@ -58,7 +58,9 @@ export function getMetadata (json) {
 export function getMdIdentification (json) {
     return json.GetRecordByIdResponse?.MD_Metadata?.identificationInfo?.MD_DataIdentification
         ||
-        json.GetRecordByIdResponse?.MD_Metadata?.identificationInfo?.SV_ServiceIdentification;
+        json.GetRecordByIdResponse?.MD_Metadata?.identificationInfo?.SV_ServiceIdentification
+        ||
+        json.MD_Metadata?.identificationInfo?.MD_DataIdentification;
 }
 
 /**
@@ -120,7 +122,7 @@ function parseFrequenzy (json) {
  * @param {String} dateType - the type of the date (e.g. publication)
  * @returns {String|undefined} formatted date
  */
-function parseDate (json, dateType) {
+export function parseDate (json, dateType) {
     const dates = getMdIdentification(json)?.citation?.CI_Citation?.date;
     let dateValue;